Taking first go at the property estimator layout
Many functions are still in abstract form, so only API purposes.
`thermoml_parser` picked up and adapted from `thermopyl`

### Major wall to overcome
**Determine how to extract Compound/Mixture info from ThermoML**

This is actually a really difficult task since the composition can be defined in several locations:
* As a `Variable` (controlled quantity that is changing)
* As a `Constraint` (constant)
* As a `Property` (measured quantity)

Within any `PureOrMixtureProperty` which is the parent node of all three of the previous ones, the compounds which participate in the measurement are listed at `Compound`. However, the following can all be mutually true across all three sub-nodes:
* No quantities need be defined, in which case it is of pure fluid (easy)
* Quantities come in all types (volume, mass, mole)
* The quantities of all compounds after considerations in `Variable`, `Constraint`, `Property` need not sum to 1
* Some components can be dependent of each other (e.g. B = 1-A)
* Subsets of components can be added together to make 1 (e.g. A+B=1 because this is the solvent that C is dissolved into)

Because these rules all seem to be contextually dependent, it becomes much harder to figure out what the correct composition of the measurement should be.

Right now I have the some of the logic in place to handle this, and it should work for pure fluids, but I have not tested this.