First pass at updating developer docs
Note that I removed information about ToolkitRegistry's from_*
methods, as it seemed to be out of date. These methods no longer
exist on ToolkitRegistry, and have been moved to Molecule and
FrozenMolecule. As the info was simply that these methods are
essentially classmethods for these latter classes, I removed it.