Performance improvements in from_openeye
OpenMM quantities are generally as quick and lightweight as
corresponding NumPy arrays, but Pint units tend to be slower. Most of
these changes simply defer unit tagging until a final step, carrying
values with implicit units in some internal functions where the units
are known and can be expected not to vary.