Heuristically check number of atoms in are_isomorphic()
Molecule.are_isomorphic() begins by comparing the
Hill formulae of each molecule, which is quite
slow especially for large molecules like proteins.
Adding a simple check for the number of atoms
first is much faster. This change cuts in half the
time taken to call from_openmm on a topology of a
protein and about 20k waters (~2min -> ~1min)